+++ src/Ably.ConsoleTest/Ably.ConsoleTest.csproj
@@ -51,6 +51,7 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Rest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
